+++ src/main/java/pl/asie/computronics/integration/enderio/DriverPowerMonitor.java
@@ -33,44 +33,77 @@ public class DriverPowerMonitor {
 				return 4;
 			}
 
-			@Callback(doc="function():number; Returns the energy currently in the conduit network")
+			@Callback(doc = "function():number; Returns the energy currently in the conduit network")
 			public Object[] getPowerInConduits(Context c, Arguments a) {
 				return new Object[] { tile.getPowerInConduits() };
 			}
 
-			@Callback(doc="function():number; Returns the max energy that can be in the conduit network")
+			@Callback(doc = "function():number; Returns the max energy that can be in the conduit network")
 			public Object[] getMaxPowerInCoduits(Context c, Arguments a) {
-				return new Object[] { tile.getMaxPowerInCoduits() };
+				return new Object[] { tile.getMaxPowerInConduits() };
 			}
 
-			@Callback(doc="function():number; Returns the energy currently in connected Capacitor Banks")
+			@Callback(doc = "function():number; Returns the energy currently in connected Capacitor Banks")
 			public Object[] getPowerInCapBanks(Context c, Arguments a) {
 				return new Object[] { tile.getPowerInCapBanks() };
 			}
 
-			@Callback(doc="function():number; Returns the max energy that can be in connected Capacitor Banks")
+			@Callback(doc = "function():number; Returns the max energy that can be in connected Capacitor Banks")
 			public Object[] getMaxPowerInCapBanks(Context c, Arguments a) {
 				return new Object[] { tile.getMaxPowerInCapBanks() };
 			}
 
-			@Callback(doc="function():number; Returns the energy currently in connected Machines")
+			@Callback(doc = "function():number; Returns the energy currently in connected Machines")
 			public Object[] getPowerInMachines(Context c, Arguments a) {
 				return new Object[] { tile.getPowerInMachines() };
 			}
 
-			@Callback(doc="function():number; Returns the max energy that can be in connected Machines")
+			@Callback(doc = "function():number; Returns the max energy that can be in connected Machines")
 			public Object[] getMaxPowerInMachines(Context c, Arguments a) {
 				return new Object[] { tile.getMaxPowerInMachines() };
 			}
 
-			@Callback(doc="function():number; Returns the average energy sent")
+			@Callback(doc = "function():number; Returns the average energy sent")
 			public Object[] getAverageEnergySent(Context c, Arguments a) {
 				return new Object[] { tile.getAveRfSent() };
 			}
 
-			@Callback(doc="function():number; Returns the average energy received")
+			@Callback(doc = "function():number; Returns the average energy received")
 			public Object[] getAverageEnergyReceived(Context c, Arguments a) {
-				return new Object[] { tile.getAveRfRecieved() };
+				return new Object[] { tile.getAveRfReceived() };
+			}
+
+			@Callback(doc = "function():boolean; Returns whether Engine Control is enabled")
+			public Object[] isEngineControlEnabled(Context c, Arguments a) {
+				return new Object[] { tile.isEngineControlEnabled() };
+			}
+
+			@Callback(doc = "function(control:boolean); Sets whether Engine Control is enabled")
+			public Object[] setEngineControlEnabled(Context c, Arguments a) {
+				tile.setEngineControlEnabled(a.checkBoolean(0));
+				return new Object[] { };
+			}
+
+			@Callback(doc = "function():number; Returns the level at which the monitor should start emitting redstone")
+			public Object[] getStartLevel(Context c, Arguments a) {
+				return new Object[] { tile.getStartLevel() };
+			}
+
+			@Callback(doc = "function(level:number); Sets the level at which the monitor should start emitting redstone")
+			public Object[] setStartLevel(Context c, Arguments a) {
+				tile.setStartLevel((float) a.checkDouble(0));
+				return new Object[] { };
+			}
+
+			@Callback(doc = "function():number; Returns the level at which the monitor should stop emitting redstone")
+			public Object[] getStopLevel(Context c, Arguments a) {
+				return new Object[] { tile.getStopLevel() };
+			}
+
+			@Callback(doc = "function(level:number); Sets the level at which the monitor should stop emitting redstone")
+			public Object[] setStopLevel(Context c, Arguments a) {
+				tile.setStopLevel((float) a.checkDouble(0));
+				return new Object[] { };
 			}
 		}
 
@@ -111,7 +144,8 @@ public class DriverPowerMonitor {
 		@Override
 		public String[] getMethodNames() {
 			return new String[] { "getPowerInConduits", "getMaxPowerInCoduits", "getPowerInCapBanks", "getMaxPowerInCapBanks",
-				"getPowerInMachines", "getMaxPowerInMachines", "getAverageEnergySent", "getAverageEnergyReceived" };
+				"getPowerInMachines", "getMaxPowerInMachines", "getAverageEnergySent", "getAverageEnergyReceived",
+				"isEngineControlEnabled", "setEngineControlEnabled", "getStartLevel", "setStartLevel", "getStopLevel", "setStopLevel" };
 		}
 
 		@Override
@@ -121,7 +155,7 @@ public class DriverPowerMonitor {
 					return new Object[] { tile.getPowerInConduits() };
 				}
 				case 1:{
-					return new Object[] { tile.getMaxPowerInCoduits() };
+					return new Object[] { tile.getMaxPowerInConduits() };
 				}
 				case 2:{
 					return new Object[] { tile.getPowerInCapBanks() };
@@ -139,7 +173,37 @@ public class DriverPowerMonitor {
 					return new Object[] { tile.getAveRfSent() };
 				}
 				case 7:{
-					return new Object[] { tile.getAveRfRecieved() };
+					return new Object[] { tile.getAveRfReceived() };
+				}
+				case 8:{
+					return new Object[] { tile.isEngineControlEnabled() };
+				}
+				case 9:{
+					if(arguments.length < 1 || !(arguments[0] instanceof Boolean)) {
+						throw new LuaException("first argument needs to be a number");
+					}
+					tile.setEngineControlEnabled((Boolean) arguments[0]);
+					return new Object[] { };
+				}
+				case 10:{
+					return new Object[] { tile.getStartLevel() };
+				}
+				case 11:{
+					if(arguments.length < 1 || !(arguments[0] instanceof Number)) {
+						throw new LuaException("first argument needs to be a number");
+					}
+					tile.setStartLevel(((Number) arguments[0]).floatValue());
+					return new Object[] { };
+				}
+				case 12:{
+					return new Object[] { tile.getStopLevel() };
+				}
+				case 13:{
+					if(arguments.length < 1 || !(arguments[0] instanceof Number)) {
+						throw new LuaException("first argument needs to be a number");
+					}
+					tile.setStopLevel(((Number) arguments[0]).floatValue());
+					return new Object[] { };
 				}
 			}
 			return null;